feat: create thumbnails + fix whiteboard name issue on mobile
