+++ .example.env
@@ -4,3 +4,5 @@ DATABASE_URL=""
 APP_DOMAIN_NAME=""
 # set a title for the sismo app creations
 APP_TITLE=""
+# set the domain of the website
+WEBSITE_DOMAIN=""
