+++ package.json
@@ -20,6 +20,7 @@
     "@types/node": "20.3.3",
     "@types/react": "18.2.14",
     "@types/react-dom": "18.2.6",
+    "@vercel/og": "^0.5.20",
     "autoprefixer": "10.4.14",
     "axios": "^1.5.0",
     "bufferutil": "^4.0.7",
