+++ src/components/header/header.css
@@ -1,6 +1,9 @@
 .header {
   display: grid;
-  grid-template-columns: 1fr 1fr 1fr;
+  /* grid-template-columns: 1fr 1fr 1fr;
+  grid-auto-columns: max-content;
+  justify-content: space-between; */
+  width: 100%;
   padding: 10px;
   background-color: #000000;
   justify-items: center;
