+++ src/components/whiteboard/whiteboard.css
@@ -1,4 +1,4 @@
-/* .whiteboard {
+.whiteboard {
   position: relative;
   display: flex;
   flex-direction: column;
@@ -14,4 +14,4 @@
   height: 100%;
   width: 100%;
   background-color: white;
-} */
+}
